refactor: move spawn to utils
