+++ packages/@guidesmiths/cuckoojs-cli/src/lib/runners/generic.runner.ts
@@ -1,47 +1,15 @@
-import type {SpawnOptions, StdioOptions} from 'child_process';
-import {spawn} from 'child_process';
-
-interface RunObject {
-	command: string;
-	args: string [];
-	cwd?: string;
-	stdio?: StdioOptions;
-}
+import spawnAsync, {RunObject} from '../utils/spawnAsync';
 
 export class GenericRunner {
-	constructor(protected binary: string) {}
+  constructor(protected binary: string) {
+  }
 
-	protected async run({
-		command,
-		args = [],
-		cwd = process.cwd(),
-		stdio = 'inherit',
-	}: RunObject): Promise<string | void> {
-		const options: SpawnOptions = {
-			cwd,
-			stdio,
-			shell: true,
-		};
-
-		return new Promise((resolve, reject) => {
-			const child = spawn(
-				this.binary,
-				args = [command, ...args],
-				options,
-			);
-
-			child.on('error', error => {
-				reject(new Error(`Child process filed with error: ${error.message}`));
-			});
-
-			child.on('close', code => {
-				if (code === 0) {
-					resolve();
-					return;
-				}
-
-				reject(new Error(`Failed to execute command: ${this.binary} ${args.join(' ')}`));
-			});
-		});
-	}
+  protected run({
+    command,
+    args = [],
+    cwd = process.cwd(),
+    stdio = 'inherit',
+  }: RunObject) {
+    return spawnAsync({command: this.binary, args: [command, ...args], stdio, cwd});
+  }
 }
